membership/web/auth.py
from config.auth_config import JWT_SECRET, JWT_CLIENT_ID, ADMIN_CLIENT_ID, ADMIN_CLIENT_SECRET, \
    AUTH_CONNECTION, AUTH_URL
from datetime import datetime, timedelta
from functools import wraps
from flask import request, Response, jsonify
import jwt
import logging
import membership
from membership.database.base import Session
from membership.database.models import Member
from membership.util.email import send_emails
import pkg_resources
import random
import requests
import string

PASSWORD_CHARS = string.ascii_letters + string.digits

def check_auth(username: str, password: str) -> bool:
    """This function is called to check if a username /
    password combination is valid.
    """
    # TODO(jesse) Switch to auth0 and don't hard code password
    return username == 'admin' and password == 'htlaeherpmes'


def deny() -> Response:
    """Sends a 401 response that enables basic auth"""
    response = jsonify({
        'status': 'error',
        'err': 'Could not verify your access level for that URL.\n'
               'You have to login with proper credentials'
    })
    response.status_code = 401
    return response


def requires_auth(admin=False):
    """ This defines a decorator which when added to a route function in flask requires authorization to
    view the route.
    """
    def decorator(f):
        @wraps(f)
        def decorated(*args, **kwargs):
            auth = request.headers.get('authorization')
            if not auth:
                return deny()
            token = auth.split()[1]
            try:
                token = jwt.decode(token, JWT_SECRET, audience=JWT_CLIENT_ID)
            except Exception as e:
                return deny()
            email = token.get('email')
            session = Session()
            try:
                member = session.query(Member).filter_by(email_address=email).one()
                authenticated = False
                if admin:
                    for role in member.roles:
                        if role.committee_id is None and role.role == 'admin':
                            authenticated = True
                if authenticated:
                    return f(*args, **kwargs)
                return deny()
            finally:
                session.close()

        return decorated
    return decorator


current_token = {}


def get_auth0_token():
    if not current_token or datetime.now() > current_token['expiry']:
        current_token.update(generate_auth0_token())
    return current_token['token']


def generate_auth0_token():
    payload = {'grant_type': "client_credentials",
               'client_id': ADMIN_CLIENT_ID,
               'client_secret': ADMIN_CLIENT_SECRET,
               'audience': AUTH_URL + 'api/v2/'}
    response = requests.post(AUTH_URL + 'oauth/token', json=payload).json()
    return {'token': response['access_token'],
            'expiry': datetime.now() + timedelta(seconds=response['expires_in'])}


def create_auth0_user(email, name):
    # create the user
    payload = {
        'connection': AUTH_CONNECTION,
        'email': email,
        'password': ''.join(random.SystemRandom().choice(PASSWORD_CHARS) for _ in range(12)),
        'user_metadata': {},
        'email_verified': False,
        'verify_email': False
    }
    headers = {'Authorization': 'Bearer ' + get_auth0_token()}
    r = requests.post(AUTH_URL + 'api/v2/users', json=payload, headers=headers)
    if r.status_code > 299:
        logging.error(r.json())
        raise Exception('Failed to create user')
    user_id = r.json()['user_id']

    # get a password change URL
    payload = {
        'result_url': 'http:localhost:3000/',
        'user_id': user_id
    }
    r = requests.post(AUTH_URL + 'api/v2/tickets/password-change', json=payload, headers=headers)
    if r.status_code > 299:
        logging.error(r.json())
        raise Exception('Failed to get password url')
    reset_url = r.json()['ticket']

    # get email verification link
    payload = {
        'result_url': reset_url,
        'user_id': user_id
    }
    r = requests.post(AUTH_URL + 'api/v2/tickets/email-verification', json=payload, headers=headers)
    if r.status_code > 299:
        logging.error(r.json())
        raise Exception('Failed to get verify url')
    validate_url = r.json()['ticket']
    # now send email with this link
    template = pkg_resources.resource_string(membership.__name__, 'templates/welcome_email.html')
    recipient_variables = {email: {'name': name, 'link': validate_url}}
    send_emails('Welcome %recipient.name%', template, recipient_variables)
