requirements.txt
alembic==0.9.2
Flask==0.12.2
Flask-Cors==3.0.2
mysqlclient==1.3.10
PyJWT==1.5.0
requests==2.17.3
SQLAlchemy==1.1.10
